Re-factor code
