+++ src/components/FormField.js
@@ -1,8 +1,9 @@
 import "bulma/css/bulma.css"
+import clsx from "clsx"
 import "./FormField.css"
 
-const FormField = ({ label, type, placeholder }) => (
-  <div className="field mt-2">
+const FormField = ({ label, type, placeholder, className }) => (
+  <div className={clsx(className, "field")}>
     <label className="label">{label}</label>
     <div className="control">
       <input className="input" type={type} placeholder={placeholder} />
